+++ src/client/games/slashchain/hand_tile.tsx
@@ -2,7 +2,7 @@ import React, { FC, useCallback, useState } from "react";
 import styled from "styled-components";
 import { Tile } from "@/games/slashchain";
 import { TileComponent } from "./tile";
-import { PickedHand } from "./my_hands";
+import { PickedHand } from "./my_field";
 import { CellComponent } from "./cell";
 
 export type HandState = "picked" | "pickable" | "disabled" | "fixed";
